taxon list: Move some source concerns out of formatter onto source.
